Support the Kubernetes aggregation layer
This patch adds support for the Kubrernetes aggregation layer. This
patch also fixes the broken e2e conformance test discussed at
https://github.com/vmware/simple-k8s-test-env/issues/10.

Many thanks to liggitt for helping to understand the reason for the
test failure and pointing to http://bit.ly/2tM4DDg for how to configure
the aggregation layer.