e2e: Fixes aws describe error when arns >20
This patch fixes an issue in the "destroy.sh" script where AWS resources
would fail on deletion because "aws elbv2 describe-tags --resource-arns"
disallows more than 20 arguments to the "--resource-arns" flag. Because
there may be more than 20 resources for which to get tags, it was
necessary to refactor "get_arns_from_tag_descriptions" to get tags in
batches, up to 20 at a time.